+++ CHANGELOG.md
@@ -10,8 +10,12 @@ the [PEP 440 version scheme](https://peps.python.org/pep-0440/#version-scheme).
 ### Added
 - py.typed
 
+### Changed
+- AbstractPlugin to NamedPluginMixin
+
 ### Fixed
 - The Alpha status classifier
+- MultiprocessingPlugin's violation of the Liskov Substitution Principle
 
 ## [v0.1.0] - 2022-12-10
 ### Added